chore: add todo for problem remaind in refactor
