+++ pp/peers/network.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"net"
 
-	"github.com/stratosnet/sds/framework/core"
 	"github.com/stratosnet/sds/pp/setting"
 	"github.com/stratosnet/sds/utils"
 )
@@ -46,19 +45,4 @@ func getInternal() string {
 	return ""
 }
 
-// PPServer
-type PPServer struct {
-	*core.Server
-}
-
-var ppServ *PPServer
-
-// GetPPServer
-func GetPPServer() *PPServer {
-	return ppServ
-}
-
-func SetPPServer(pp *PPServer) {
-	ppServ = pp
-}
